backend/lenta/requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4
djoser==2.2.0
gunicorn==21.2.0
djangorestframework==3.14.0
python-dotenv
progress==1.6
drf-spectacular
drf-spectacular-sidecar
black==23.7.0
ruff==0.0.285
psycopg2-binary==2.9.6
openpyxl==3.1.2
django-filter==23.3